refactor: carReader, carWriter 추상화 클래스 import / updateCar 내부 유효성 검사 로직 Entity로 이동
