+++ back/src/main/java/com/example/_thecore_back/car/domain/CarEntity.java
@@ -1,5 +1,7 @@
 package com.example._thecore_back.car.domain;
 
+import com.example._thecore_back.car.controller.dto.CarDetailDto;
+import com.example._thecore_back.car.controller.dto.CarRequestDto;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -38,5 +40,34 @@ public class CarEntity {
 
     private Integer emulatorId; // 연결된 애뮬레이터 아이디
 
+    public void updateInfo(CarRequestDto carRequest) {
+        if (carRequest.getBrand() != null && !carRequest.getBrand().isBlank()) {
+            setBrand(carRequest.getBrand());
+        }
 
+        if(carRequest.getModel() != null && !carRequest.getModel().isBlank()) {
+            setModel(carRequest.getModel());
+        }
+
+        if(carRequest.getCarYear() != null && !carRequest.getCarYear().equals(0)) {
+            setCarYear(carRequest.getCarYear());
+        }
+
+        if(carRequest.getStatus() != null && !carRequest.getStatus().isBlank()) {
+            var status = CarStatus.fromDisplayName(carRequest.getStatus());
+            setStatus(status);
+        }
+
+        if(carRequest.getCarType() != null && !carRequest.getCarType().isBlank()) {
+            setCarType(carRequest.getCarType());
+        }
+
+        if(carRequest.getCarNumber() != null && !carRequest.getCarNumber().isBlank()) {
+            setCarNumber(carRequest.getCarNumber());
+        }
+
+        if (carRequest.getSumDist() != null && carRequest.getSumDist() >= 0) {
+            setSumDist(carRequest.getSumDist());
+        }
+    }
 }
